DFSB Overhaul
Overhauled the DFSB generator to be a bit more modular. Changes were
made to implement a way to flag the generator to attempt to use the last
word on the previous line to generate the next line. Working on
implmentation that will generate the whole Haiku from one start node.